refactor: move service/model to model/service package
- AvailableSlot, SlotQuery, TimeRange 패키지 이동
  service.model → model.service
- 관련 import 일괄 업데이트
  SlotCalculationService, ClosureRescheduleService, SlotController, SlotResponse, SlotCalculationServiceTest, TimeRangeTest
