+++ appointment-core/src/main/kotlin/io/bluetape4k/clinic/appointment/model/service/SlotQuery.kt
@@ -1,4 +1,4 @@
-package io.bluetape4k.clinic.appointment.service.model
+package io.bluetape4k.clinic.appointment.model.service
 
 import java.time.LocalDate
